Rename `core.js` to `index.js`
