+++ packages/kryo-json/src/lib/json-stream-writer.ts
@@ -2,7 +2,7 @@
  * @module kryo/writers/json-stream
  */
 
-import { Writer } from "kryo/lib/core.js";
+import { Writer } from "kryo";
 
 export interface WritableStream {
   /**
