+++ packages/kryo-json/src/lib/json-value-reader.ts
@@ -3,7 +3,7 @@
  */
 
 import incident from "incident";
-import { Reader, ReadVisitor } from "kryo/lib/core.js";
+import { Reader, ReadVisitor } from "kryo";
 import { createInvalidTypeError } from "kryo/lib/errors/invalid-type.js";
 
 import { JsonValue } from "./json-value.js";
