+++ packages/kryo-qs/src/lib/qs-value-reader.ts
@@ -3,8 +3,8 @@
  */
 
 import incident from "incident";
+import { Reader, ReadVisitor } from "kryo";
 import { JsonReader } from "kryo-json/lib/json-reader.js";
-import { Reader, ReadVisitor } from "kryo/lib/core.js";
 import { createInvalidTypeError } from "kryo/lib/errors/invalid-type.js";
 
 export class QsValueReader implements Reader<any> {
