+++ packages/kryo/src/lib/_helpers/test-error.ts
@@ -1,5 +1,5 @@
-import { Type } from "../core.js";
 import { createInvalidValueError } from "../errors/invalid-value.js";
+import { Type } from "../index.js";
 
 /**
  * Calls the `.testError` method of `type` or falls back to an implementation derived from `.test`.
