+++ packages/kryo/src/lib/array.ts
@@ -2,12 +2,12 @@ import incident from "incident";
 
 import { lazyProperties } from "./_helpers/lazy-properties.js";
 import { testError } from "./_helpers/test-error.js";
-import { IoType, Lazy, Reader, Type, Writer } from "./core.js";
 import { createInvalidArrayItemsError } from "./errors/invalid-array-items.js";
 import { createInvalidTypeError } from "./errors/invalid-type.js";
 import { createLazyOptionsError } from "./errors/lazy-options.js";
 import { createMaxArrayLengthError } from "./errors/max-array-length.js";
 import { createMinArrayLengthError } from "./errors/min-array-length.js";
+import { IoType, Lazy, Reader, Type, Writer } from "./index.js";
 import { readVisitor } from "./readers/read-visitor.js";
 
 export type Name = "array";
