+++ packages/kryo/src/lib/date.ts
@@ -1,6 +1,6 @@
-import { IoType, Ord, Reader, VersionedType, Writer } from "./core.js";
 import { createInvalidTimestampError } from "./errors/invalid-timestamp.js";
 import { createInvalidTypeError } from "./errors/invalid-type.js";
+import { IoType, Ord, Reader, VersionedType, Writer } from "./index.js";
 import { readVisitor } from "./readers/read-visitor.js";
 
 export type Name = "date";
