+++ packages/kryo/src/lib/integer.ts
@@ -1,10 +1,10 @@
 import incident from "incident";
 
 import { lazyProperties } from "./_helpers/lazy-properties.js";
-import { IoType, Lazy, Ord, Reader, VersionedType, Writer } from "./core.js";
 import { createInvalidIntegerError } from "./errors/invalid-integer.js";
 import { createInvalidTypeError } from "./errors/invalid-type.js";
 import { createLazyOptionsError } from "./errors/lazy-options.js";
+import { IoType, Lazy, Ord, Reader, VersionedType, Writer } from "./index.js";
 import { readVisitor } from "./readers/read-visitor.js";
 
 export type Name = "integer";
