+++ packages/kryo/src/lib/literal.ts
@@ -2,8 +2,8 @@ import incident from "incident";
 
 import { lazyProperties } from "./_helpers/lazy-properties.js";
 import { testError } from "./_helpers/test-error.js";
-import { IoType, Lazy, Reader, Type, Writer } from "./core.js";
 import { createLazyOptionsError } from "./errors/lazy-options.js";
+import { IoType, Lazy, Reader, Type, Writer } from "./index.js";
 
 export type Name = "literal";
 export const name: Name = "literal";
