+++ packages/kryo/src/lib/map.ts
@@ -2,10 +2,10 @@ import incident from "incident";
 
 import { lazyProperties } from "./_helpers/lazy-properties.js";
 import { testError } from "./_helpers/test-error.js";
-import { IoType, Lazy, Reader, VersionedType, Writer } from "./core.js";
 import { createInvalidTypeError } from "./errors/invalid-type.js";
 import { createLazyOptionsError } from "./errors/lazy-options.js";
 import { createNotImplementedError } from "./errors/not-implemented.js";
+import { IoType, Lazy, Reader, VersionedType, Writer } from "./index.js";
 import { readVisitor } from "./readers/read-visitor.js";
 
 export type Name = "map";
