+++ packages/kryo/src/lib/try-union.ts
@@ -2,8 +2,8 @@ import incident from "incident";
 
 import { lazyProperties } from "./_helpers/lazy-properties.js";
 import { testError } from "./_helpers/test-error.js";
-import { IoType, Lazy, Reader, Type, VersionedType, Writer } from "./core.js";
 import { createLazyOptionsError } from "./errors/lazy-options.js";
+import { IoType, Lazy, Reader, Type, VersionedType, Writer } from "./index.js";
 
 export type Name = "union";
 export const name: Name = "union";
