+++ packages/kryo/src/lib/ts-enum.ts
@@ -2,9 +2,9 @@ import incident from "incident";
 
 import { rename } from "./_helpers/case-style.js";
 import { lazyProperties } from "./_helpers/lazy-properties.js";
-import { CaseStyle, IoType, Lazy, Reader, Writer } from "./core.js";
 import { createLazyOptionsError } from "./errors/lazy-options.js";
 import { createNotImplementedError } from "./errors/not-implemented.js";
+import { CaseStyle, IoType, Lazy, Reader, Writer } from "./index.js";
 import { readVisitor } from "./readers/read-visitor.js";
 
 /**
