+++ packages/kryo/src/lib/ucs2-string.ts
@@ -1,7 +1,6 @@
 import incident from "incident";
 
 import { lazyProperties } from "./_helpers/lazy-properties.js";
-import { IoType, Lazy, Reader, VersionedType, Writer } from "./core.js";
 import { createInvalidTypeError } from "./errors/invalid-type.js";
 import { createLazyOptionsError } from "./errors/lazy-options.js";
 import { createLowerCaseError } from "./errors/lower-case.js";
@@ -9,6 +8,7 @@ import { createMaxUcs2StringLengthError } from "./errors/max-ucs2-string-length.
 import { createMinUcs2StringLengthError } from "./errors/min-ucs2-string-length.js";
 import { createNotTrimmedError } from "./errors/not-trimmed.js";
 import { createPatternNotMatchedError } from "./errors/pattern-not-matched.js";
+import { IoType, Lazy, Reader, VersionedType, Writer } from "./index.js";
 import { readVisitor } from "./readers/read-visitor.js";
 
 export type Name = "ucs2-string";
