+++ packages/kryo/src/test/types/custom.spec.ts
@@ -1,9 +1,9 @@
 import BSON from "bson";
 import incident from "incident";
 
-import { Reader, Writer } from "../../lib/core.js";
 import { CustomType } from "../../lib/custom.js";
 import { createInvalidTypeError } from "../../lib/errors/invalid-type.js";
+import { Reader, Writer } from "../../lib/index.js";
 import { readVisitor } from "../../lib/readers/read-visitor.js";
 import { runTests, TypedValue } from "../helpers/test.js";
