+++ packages/kryo/src/test/types/tagged-union.spec.ts
@@ -1,6 +1,6 @@
 import BSON from "bson";
 
-import { CaseStyle } from "../../lib/core.js";
+import { CaseStyle } from "../../lib/index.js";
 import { IntegerType } from "../../lib/integer.js";
 import { LiteralType } from "../../lib/literal.js";
 import { RecordType } from "../../lib/record.js";
